refactor integrator to enum
