+++ src/renderer.rs
@@ -5,11 +5,9 @@ use crate::samplers::{JitteredSampler, UniformSampler};
 use crate::tracer::scene::Scene;
 use crate::tracer::camera::Camera;
 #[allow(unused_imports)]
-use crate::tracer::integrators::{Integrator, DirectLightingIntegrator, PathTracingIntegrator};
+use crate::tracer::integrator::Integrator;
 
 type PxSampler = JitteredSampler;
-//type Intgrtr = DirectLightingIntegrator;
-type Intgrtr = PathTracingIntegrator;
 
 pub fn _render(
     img_height: usize,
@@ -18,9 +16,9 @@ pub fn _render(
     px_width: f64,
     num_samples: usize,
     cam: Camera,
-    scene: Scene,
+    scene: &Scene,
 ) -> Vec<DVec3> {
-    let integrator = Intgrtr::new(scene);
+    let integrator = Integrator::PathTrace;
     (0..img_height).into_par_iter().flat_map(|y: usize| {
         (0..img_width).map(|x: usize| {
             let u = x as f64 * px_width;
@@ -28,8 +26,8 @@ pub fn _render(
 
             PxSampler::new(num_samples).map(|rand_sq: DVec2| {
                 integrator.integrate(
+                    scene,
                     &cam.ray_at(u + rand_sq.x*px_width, v + rand_sq.y*px_height),
-                    0
                 )
             }).fold(DVec3::ZERO, |acc: DVec3, c: DVec3| acc + c)
                 / num_samples as f64
